requirements.txt
openai==1.37.1
tika==2.6.0
streamlit==1.36.0
tiktoken==0.7.0
python-dotenv
natsort==8.4.0
chromadb==0.5.5
pandas==2.2.2
numpy==1.26.4